Speed up group by string (domains)
Doing groupby string column (domains in reverse order) is one of our worse
performing queries we have. Mostly because we're grouping by 10s and 100s
thousands of unique values. I've beeen trying to make this peform better for
a long time. We got a nice but small speedup (5%) before by using CityHash64.

After spending more time with perf, it looked like of the if condition branches
was always taking a lot of time. For some reason it kept getting miss
predicted. It was a simple boolean condition.

On ver large groypby (100,000 elements) this improves execution time by 5% to
10%. No observable regression to small inputs once we introducted the hash size
check.